feat: reorganize old code
